+++ nx/blocks/form/views/editor.js
